-: Fixes to make scrutinizer happy
