+++ src/Server/Config/Matchers.php
@@ -25,4 +25,21 @@ class Matchers
     const SAME_STRING = 'isSameString';
     const CONTAINS = 'contains';
     const SAME_JSON = 'isSameJsonObject';
+
+    const VALID_MATCHERS = [
+        self::CONTAINS,
+        self::EQUAL_TO,
+        self::MATCHES,
+        self::SAME_JSON,
+        self::SAME_STRING,
+    ];
+
+    /**
+     * @param string $matcherName
+     * @return bool
+     */
+    public static function isValidMatcher($matcherName)
+    {
+        return in_array($matcherName, self::VALID_MATCHERS);
+    }
 }
