+++ src/Server/Http/RequestFilters/ConvertToCondition.php
@@ -21,9 +21,10 @@ namespace Mcustiel\Phiremock\Server\Http\RequestFilters;
 use Mcustiel\Phiremock\Domain\Condition;
 use Mcustiel\Phiremock\Server\Config\Matchers;
 use Mcustiel\SimpleRequest\Exception\FilterErrorException;
+use Mcustiel\SimpleRequest\Filter\AbstractEmptySpecificationFilter;
 use Mcustiel\SimpleRequest\Interfaces\FilterInterface;
 
-class ConvertToCondition implements FilterInterface
+class ConvertToCondition extends AbstractEmptySpecificationFilter implements FilterInterface
 {
     /**
      * {@inheritdoc}
@@ -33,7 +34,7 @@ class ConvertToCondition implements FilterInterface
     public function filter($value)
     {
         if (null === $value) {
-            return;
+            return null;
         }
         $this->checkValueIsValidOrThrowException($value);
         $matcher = key($value);
@@ -41,16 +42,6 @@ class ConvertToCondition implements FilterInterface
         $this->validateValueOrThrowException($value[$matcher]);
 
         return new Condition($matcher, $value[$matcher]);
-    }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @see \Mcustiel\SimpleRequest\Interfaces\Specificable::setSpecification()
-     * @SuppressWarnings("unused")
-     */
-    public function setSpecification($specification = null)
-    {
     }
 
     /**
@@ -72,23 +63,9 @@ class ConvertToCondition implements FilterInterface
      */
     private function validateMatcherOrThrowException($matcher)
     {
-        if (!$this->isValidCondition($matcher)) {
+        if (!Matchers::isValidMatcher($matcher)) {
             throw new FilterErrorException('Invalid condition matcher specified: ' . $matcher);
         }
-    }
-
-    /**
-     * @param mixed $matcherName
-     *
-     * @return bool
-     */
-    private function isValidCondition($matcherName)
-    {
-        return Matchers::EQUAL_TO === $matcherName
-            || Matchers::MATCHES === $matcherName
-            || Matchers::SAME_STRING === $matcherName
-            || Matchers::CONTAINS === $matcherName
-            || Matchers::SAME_JSON === $matcherName;
     }
 
     /**
